refactor(admin): extract listDatasetsWithEsWarnings shared helper
